fix: make subprocess exit set started state
If an `ipfsd-daemon` instance connects to a running API it does not have a
`this.subprocess` property, yet if we try to stop the node it tries to access
that property, leading to this sort of error:

```
  2) DelegatedPeerRouting
       "after all" hook in "DelegatedPeerRouting":
     TypeError: Cannot read property 'stderr' of null
      at Daemon.stop (node_modules/ipfsd-ctl/src/ipfsd-daemon.js:260:21)
      at processTicksAndRejections (node:internal/process/task_queues:93:5)
      at async Promise.all (index 0)
```

This PR moves `this.subprocess` access to where it is created and lets the
process exiting set `this.started` to `false` as otherwise the state can
get out of sync with reality.

+++ src/ipfsd-daemon.js
@@ -10,6 +10,7 @@ const path = require('path')
 const os = require('os')
 const tempWrite = require('temp-write')
 const { checkForRunningApi, repoExists, tmpDir, defaultRepo } = require('./utils')
+const waitFor = require('p-wait-for')
 
 const daemonLog = {
   info: debug('ipfsd-ctl:daemon:stdout'),
@@ -157,33 +158,41 @@ class Daemon {
    * @returns {Promise<Daemon>}
    */
   async start () {
-    const args = ['daemon']
-    const opts = this.opts.ipfsOptions
-    // add custom args
-    args.push(...this.opts.args)
-
-    if (opts.pass && this.opts.type === 'js') {
-      args.push('--pass', '"' + opts.pass + '"')
-    }
-    if (opts.offline) {
-      args.push('--offline')
-    }
-    if (opts.preload && this.opts.type === 'js') {
-      args.push('--enable-preload', Boolean(opts.preload.enabled))
-    }
-    if (opts.EXPERIMENTAL && opts.EXPERIMENTAL.sharding && this.opts.type === 'js') {
-      args.push('--enable-sharding-experiment')
-    }
-    if (opts.EXPERIMENTAL && opts.EXPERIMENTAL.ipnsPubsub) {
-      args.push('--enable-namesys-pubsub')
-    }
-
     // Check if a daemon is already running
     const api = checkForRunningApi(this.path)
+
     if (api) {
       this._setApi(api)
+    } else if (!this.exec) {
+      throw new Error('No executable specified')
     } else {
+      const args = ['daemon']
+      const opts = this.opts.ipfsOptions
+      // add custom args
+      args.push(...this.opts.args)
+
+      if (opts.pass && this.opts.type === 'js') {
+        args.push('--pass', '"' + opts.pass + '"')
+      }
+
+      if (opts.offline) {
+        args.push('--offline')
+      }
+
+      if (opts.preload && this.opts.type === 'js') {
+        args.push('--enable-preload', Boolean(opts.preload.enabled))
+      }
+
+      if (opts.EXPERIMENTAL && opts.EXPERIMENTAL.sharding && this.opts.type === 'js') {
+        args.push('--enable-sharding-experiment')
+      }
+
+      if (opts.EXPERIMENTAL && opts.EXPERIMENTAL.ipnsPubsub) {
+        args.push('--enable-namesys-pubsub')
+      }
+
       let output = ''
+
       const ready = new Promise((resolve, reject) => {
         this.subprocess = execa(this.exec, args, {
           env: this.env
@@ -212,8 +221,25 @@ class Daemon {
           }
         }
         this.subprocess.stdout.on('data', readyHandler)
-        this.subprocess.catch(err => reject(translateError(err)))
+        this.subprocess.catch(err => {
+          if (this.started) {
+            this.started = false
+            reject(translateError(err))
+          }
+        })
+        this.subprocess.on('exit', () => {
+          if (this.started) {
+            this.started = false
+            this.subprocess.stderr.removeAllListeners()
+            this.subprocess.stdout.removeAllListeners()
+
+            if (this.disposable) {
+              this.cleanup().catch(() => {})
+            }
+          }
+        })
       })
+
       await ready
     }
 
@@ -228,43 +254,70 @@ class Daemon {
   /**
    * Stop the daemon.
    *
+   * @param {object} options
+   * @param {number} options.timeout - How long to wait for the daemon to stop
    * @returns {Promise<Daemon>}
    */
-  async stop () {
+  async stop (options = {}) {
+    const timeout = options.timeout || 60000
+
     if (!this.started) {
       return this
     }
 
     let killTimeout
     let killed = false
-    if (this.opts.forceKill !== false) {
-      killTimeout = setTimeout(() => {
-        // eslint-disable-next-line no-console
-        console.error(new Error(`Timeout stopping ${this.opts.type} node. Process ${this.subprocess.pid} will be force killed now.`))
-        killed = true
+    if (this.subprocess) {
+      if (this.opts.forceKill !== false) {
+        killTimeout = setTimeout(() => {
+          // eslint-disable-next-line no-console
+          console.error(new Error(`Timeout stopping ${this.opts.type} node. Process ${this.subprocess.pid} will be force killed now.`))
+          killed = true
 
-        this.subprocess.kill('SIGKILL')
-      }, this.opts.forceKillTimeout)
-    }
-
-    try {
-      await this.api.stop()
-    } catch (err) {
-      if (!killed) {
-        throw err // if was killed then ignore error
+          this.subprocess.kill('SIGKILL')
+        }, this.opts.forceKillTimeout)
       }
 
-      daemonLog.info('Daemon was force killed')
+      this.subprocess.cancel()
+
+      try {
+        await this.subprocess
+      } catch (err) {
+        if (!this.subprocess.killed) {
+          throw err
+        }
+      }
+    } else {
+      try {
+        await this.api.stop()
+      } catch (err) {
+        if (!killed) {
+          throw err // if was killed then ignore error
+        }
+
+        daemonLog.info('Daemon was force killed')
+      }
     }
 
     clearTimeout(killTimeout)
-    this.subprocess.stderr.removeAllListeners()
-    this.subprocess.stdout.removeAllListeners()
-    this.started = false
 
-    if (this.disposable) {
-      await this.cleanup()
+    if (!this.subprocess) {
+      // if we have a subprocess, this.started will be set when it exits
+      this.started = false
+    } else {
+      // wait for the subprocess to exit and declare ourselves stopped
+      await waitFor(() => !this.started, {
+        timeout
+      })
+
+      if (this.disposable) {
+        // wait for the cleanup routine to run after the subprocess has exited
+        await waitFor(() => this.clean, {
+          timeout
+        })
+      }
     }
+
     return this
   }
